pmix/ext2x: automatically generate ext2x component from pmix2x sources
